+++ packages/avatar/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.14](https://github.com/equinor/fusion-web-components/compare/@equinor/fusion-wc-avatar@1.0.13...@equinor/fusion-wc-avatar@1.0.14) (2021-09-16)
+
+**Note:** Version bump only for package @equinor/fusion-wc-avatar
+
+
+
+
+
 ## [1.0.13](https://github.com/equinor/fusion-web-components/compare/@equinor/fusion-wc-avatar@1.0.12...@equinor/fusion-wc-avatar@1.0.13) (2021-09-15)